refactor: 라우트 유틸로 코드 분리 (#5)
